Restructure macros crate.
